refactor: separated seeding data into a separate file
